Redefine termite colors to disable bright bold
Termite uses color 8 - 15 for bold text, but the solarized palette
defines additional colors in this range, which causes e.g. bold yellow
to appear as bold gray.
This patch uses the same colors for 0 - 7 and 8 - 15 to ensure the
correct bold color. Additionally, it redefines some colors in the 256
color range to enable the full solarized palette in vim.